requirements.txt
wheel
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13.0
idna==3.3
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
python-http-client==3.3.7
requests==2.27.1
selenium==4.1.5
sendgrid==6.9.7
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starkbank-ecdsa==2.0.3
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0